requirements.txt
python-dotenv==1.0.1
PyPDF2==3.0.1
google-genai==1.11.0
pydantic==2.11.3
kokoro==0.9.4
soundfile==0.13.1
pydub==0.25.1
pillow==11.2.1
librosa==0.11.0
transformers==4.48.2